refactor: Move GetStats method from stats.go to docker.go
- Consolidate Docker client methods in docker.go
- Remove separate stats.go file for better code organization
- GetStats method now part of the main Client struct
